backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
opencv-python
face-recognition
qrcode
requests
pytesseract
reportlab